backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pymysql==1.1.0
python-decouple==3.8
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
